Removed EditModal file and moved component in the Menu file
